+++ src/main/java/com/curso/angular/endpoints/interfaces/IClienteController.java
@@ -3,6 +3,6 @@ package com.curso.angular.endpoints.interfaces
 import com.curso.angular.base.controller.IControllerCrudAbstract;
 import com.curso.angular.entity.Cliente;
 
-public interface IClienteController extends IControllerCrudAbstract<Cliente, Long>{
+public interface IClienteController extends IControllerCrudAbstract<Cliente, Integer>{
 
 }
